infoprint, FDS moviefix, alignment bugfixing
git-svn-id: file:///home/notaz/opt/svn/fceu@139 be3aeb3a-fb24-0410-a615-afba39da0efa

+++ sound.h
@@ -70,7 +70,7 @@ extern int32 fhinc;
 
 uint8 FCEU_GetJoyJoy(void);
 
-int GetSoundBuffer(int32 **W);
+int GetSoundBuffer(int16 **W);
 int FlushEmulateSound(void);
 extern uint32 Wave[2048];
 extern int32 WaveFinal[2048];
